chore: 투수페이지 sever client 분리
